fixed bug with update caching, moved everything to SchemaManager
